// templates/recruiterflow/types.ts
export interface RecruiterFlowLocationResponse {
    id: number;
    name: string;
    city: string | null;
    country: string | null;
    details: string | null;
    iso_3166_1_alpha_2_code: string | null;
    location_type: string;
    location_type_id: number;
    postal_code: string | null;
    state: string | null;
    zipcode: string | null;
}